// generated/model/manedUtbetaling.ts
/**
 * Generated by orval v6.16.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ManedUtbetaling {
    tittel: string;
    belop: number;
    utbetalingsdato?: string;
    forfallsdato?: string;
    status: string;
    fiksDigisosId: string;
    fom?: string;
    tom?: string;
    mottaker?: string;
    annenMottaker: boolean;
    kontonummer?: string;
    utbetalingsmetode?: string;
}